tui: Move file review browser module
The TUI file-review package contains action, display, prompt, session, and
browser code under the same package path.

The browser loop still lives in the package initializer, so the filesystem
tree cannot distinguish package identity from the file chooser and review
loop implementation.

This commit addresses that by moving the browser implementation into
git_stage_batch.tui.file_review.browser and importing the handlers from the
action dispatcher.

The following tests commit will update direct test imports and add a
package initializer guard.
